+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 - New CSG (Constant Stochastic Gradient) approximate posterior sampling
   algorithm (#2544)
 - Michael Osthege added support for population-samplers and implemented differential evolution metropolis (`DEMetropolis`).  For models with correlated dimensions that can not use gradient-based samplers, the `DEMetropolis` sampler can give higher effective sampling rates. (also see [PR#2735](https://github.com/pymc-devs/pymc3/pull/2735))
+- Forestplot supports multiple traces (#2736)
 
 ### Fixes